requirements.txt
aiogram==3.10.0
aiosqlite==0.20.0
python-dotenv==1.0.1